requirements.txt
channels
django-widget-tweaks
redis
channels_redis

